--- runtime/src/swam/internals/store/Store.scala
@@ -1,44 +1,0 @@
-/*
- * Copyright 2018 Lucas Satabin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package swam
-package internals
-package store
-
-import runtime._
-
-/** An abstract representation of the VM store.
-  */
-trait Store {
-
-  /** Returns the global value at the given address. */
-  def globalValue(addr: Address): Value
-
-  /** Sets the value to the global. If it is not mutable,
-    *  an exception is thrown.
-    */
-  def updateGlobalValue(addr: Address, v: Value): Unit
-
-  /** Returns the memory instance at the given address. */
-  def memory(addr: Address): MemoryInstance
-
-  /** Returns the function instance at the given address. */
-  def function(addr: Address): FuncInstance
-
-  /** Returns the table instance at the given address. */
-  def table(addr: Address): TableInstance
-
-}
