+++ runtime/src/swam/runtime/internals/store/FuncInstance.scala
@@ -15,20 +15,10 @@
  */
 
 package swam
+package runtime
 package internals
+package store
 
-package object interpreter {
+import java.nio.ByteBuffer
 
-  type Label = Long
-
-  implicit class LabelOps(val l: Label) extends AnyVal {
-    def arity: Int = ((l >> 32) & 0xffffffff).toInt
-    def cont: Int = (l & 0xffffffff).toInt
-  }
-
-  object Label {
-    def apply(arity: Int, cont: Int): Label =
-      (arity << 32l) | cont
-  }
-
-}
+class FuncInstance(val tpe: FuncType, val code: ByteBuffer, val locals: Array[ValType])
