+++ runtime/src/swam/runtime/internals/store/SimpleStore.scala
@@ -15,6 +15,7 @@
  */
 
 package swam
+package runtime
 package internals
 package store
